requirements.txt
fastapi
uvicorn
alembic
asyncpg
sqlalchemy
pydantic
python-jose
pyjwt